+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+Version 2.2.1 (07-05-2020)
+-----------------------------------------------------------
+
+*   Fixed memoryleak regarding websocket timeout and
+    lack of application cleanup
+*   Fixed bug related to shown exception on sendfile
+
 Version 2.2.0 (17-02-2020)
 -----------------------------------------------------------